prod_requirements.txt
Django==2.0.7
djangorestframework==3.8.2
psycopg2==2.7.3.2
gunicorn==19.9.0
django-filter==2.0.0
django-cors-headers==2.4.0
python-dateutil==2.7.3
django-request-logging==0.6.5
pyjwt==1.6.4
pycryptodome==3.6.1
